+++ client/src/main/scala/mmo/client/graphics/TileAtlas.scala
@@ -11,15 +11,15 @@ final case class TileAtlas(
   imageHeight: Int
 ) {
 
-  def render(nvg: Long, rectOnScreen: Rect[Float], positionOnTexture: V2[Int], scaleFactor: Int): Unit = {
+  def render(nvg: Long, rectOnScreen: Rect[Double], positionOnTexture: V2[Int], scaleFactor: Int): Unit = {
     // See https://github.com/memononen/nanovg/issues/348
 
-    val textureOffset = ((scaleFactor * TileAtlas.tileSize) *: positionOnTexture).map(_.toFloat)
+    val textureOffset = ((scaleFactor * TileAtlas.tileSize) *: positionOnTexture).map(_.toDouble)
     val patternPosition = rectOnScreen.xy - textureOffset
     val _ = nvgImagePattern(
       nvg,
-      patternPosition.x,
-      patternPosition.y,
+      patternPosition.x.toFloat,
+      patternPosition.y.toFloat,
       (scaleFactor * imageWidth).toFloat,
       (scaleFactor * imageHeight).toFloat,
       0, imageHandle, 1.0f, TileAtlas.paint
@@ -28,10 +28,10 @@ final case class TileAtlas(
     nvgBeginPath(nvg)
     nvgRect(
       nvg,
-      rectOnScreen.xy.x,
-      rectOnScreen.xy.y,
-      rectOnScreen.wh.x,
-      rectOnScreen.wh.y
+      rectOnScreen.xy.x.toFloat,
+      rectOnScreen.xy.y.toFloat,
+      rectOnScreen.wh.x.toFloat,
+      rectOnScreen.wh.y.toFloat
     )
     nvgFill(nvg)
   }
