+++ common/src/main/scala/mmo/common/api/Constants.scala
@@ -4,11 +4,11 @@ import mmo.common.linear.{Rect, V2}
 
 object Constants {
 
-  val playerTilePerSecond: Float = 4.0f
+  val playerTilePerSecond: Double = 4.0
 
-  val playerHitbox: Rect[Float] = Rect(
-    xy = V2(0.075f, 0.7f),
-    wh = V2(0.85f, 0.3f)
+  val playerHitbox: Rect[Double] = Rect(
+    xy = V2(0.075, 0.7),
+    wh = V2(0.85, 0.3)
   )
 
   val maxMessageBytes: Int = 4096
