+++ common/src/main/scala/mmo/common/api/Direction.scala
@@ -4,7 +4,7 @@ import mmo.common.linear.V2
 
 final case class Direction(asInt: Int) extends AnyVal {
 
-  def vector: V2[Float] = Direction.vectors(asInt)
+  def vector: V2[Double] = Direction.vectors(asInt)
   def lookDirection: LookDirection = Direction.lookDirections(asInt)
 
   def isMoving: Boolean = asInt != 0
@@ -21,16 +21,16 @@ object Direction {
   val leftUp = Direction(7)
   val rightUp = Direction(8)
 
-  private val vectors: Array[V2[Float]] = Array(
+  private val vectors: Array[V2[Double]] = Array(
     V2(0, 0),
     V2(1, 0),
     V2(0, 1),
     V2(-1, 0),
     V2(0, -1),
-    V2(1.0f, 1.0f).normalize,
-    V2(-1.0f, 1.0f).normalize,
-    V2(-1, -1.0f).normalize,
-    V2(1.0f, -1.0f).normalize
+    V2(1.0, 1.0).normalize,
+    V2(-1.0, 1.0).normalize,
+    V2(-1, -1.0).normalize,
+    V2(1.0, -1.0).normalize
   )
 
   private val lookDirections: Array[LookDirection] = Array(
